third_party/blink/web_tests/http/tests/xmlhttprequest/origin-allowlisting-ip-addresses-with-subdomains-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://127.0.0.1:8000/xmlhttprequest/resources/get.txt' from origin 'http://localhost:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to XMLHttpRequest at 'http://127.0.0.1:8000/xmlhttprequest/resources/get.txt' from origin 'http://localhost:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
Specifying that an IP address should match subdomains doesn't make sense. This test verifies that it doesn't do anything.

Testing: http://127.0.0.1:8000/xmlhttprequest/resources/get.txt (sync)
FAIL: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/xmlhttprequest/resources/get.txt'.
Testing: http://127.0.0.1:8000/xmlhttprequest/resources/get.txt (async)
FAIL: 0
